Agent: Move get_auth_options() to auth_options.py
