+++ monkey/infection_monkey/exploit/powershell.py
@@ -17,8 +17,8 @@ from infection_monkey.exploit.powershell_utils.auth_options import (
     AUTH_NEGOTIATE,
     ENCRYPTION_AUTO,
     AuthOptions,
+    get_auth_options,
 )
-from infection_monkey.exploit.powershell_utils.auth_options_generators import get_auth_options
 from infection_monkey.exploit.powershell_utils.credential_generators import get_credentials
 from infection_monkey.exploit.powershell_utils.credentials import Credentials
 from infection_monkey.exploit.powershell_utils.utils import (
