+++ monkey/infection_monkey/exploit/powershell_utils/auth_options.py
@@ -1,4 +1,7 @@
 from dataclasses import dataclass
+from typing import List
+
+from infection_monkey.exploit.powershell_utils.credentials import Credentials
 
 AUTH_BASIC = "basic"
 AUTH_NEGOTIATE = "negotiate"
@@ -11,3 +14,17 @@ class AuthOptions:
     auth_type: str
     encryption: str
     ssl: bool
+
+
+def get_auth_options(credentials: List[Credentials], use_ssl: bool) -> List[AuthOptions]:
+    auth_options = []
+
+    for creds in credentials:
+        # Passwordless login only works with SSL false, AUTH_BASIC and ENCRYPTION_NEVER
+        ssl = False if creds.password == "" else use_ssl
+        auth_type = AUTH_BASIC if creds.password == "" else AUTH_NEGOTIATE
+        encryption = ENCRYPTION_NEVER if creds.password == "" else ENCRYPTION_AUTO
+
+        auth_options.append(AuthOptions(auth_type, encryption, ssl))
+
+    return auth_options
