--- monkey/infection_monkey/exploit/powershell_utils/auth_options_generators.py
@@ -1,24 +1,0 @@
-from typing import List
-
-from infection_monkey.exploit.powershell_utils.auth_options import (
-    AUTH_BASIC,
-    AUTH_NEGOTIATE,
-    ENCRYPTION_AUTO,
-    ENCRYPTION_NEVER,
-    AuthOptions,
-)
-from infection_monkey.exploit.powershell_utils.credentials import Credentials
-
-
-def get_auth_options(credentials: List[Credentials], use_ssl: bool) -> List[AuthOptions]:
-    auth_options = []
-
-    for creds in credentials:
-        # Passwordless login only works with SSL false, AUTH_BASIC and ENCRYPTION_NEVER
-        ssl = False if creds.password == "" else use_ssl
-        auth_type = AUTH_BASIC if creds.password == "" else AUTH_NEGOTIATE
-        encryption = ENCRYPTION_NEVER if creds.password == "" else ENCRYPTION_AUTO
-
-        auth_options.append(AuthOptions(auth_type, encryption, ssl))
-
-    return auth_options
