+++ monkey/tests/unit_tests/infection_monkey/exploit/powershell_utils/test_auth_options.py
@@ -4,8 +4,8 @@ from infection_monkey.exploit.powershell_utils.auth_options import (
     AUTH_NEGOTIATE,
     ENCRYPTION_AUTO,
     ENCRYPTION_NEVER,
+    get_auth_options,
 )
-from infection_monkey.exploit.powershell_utils.auth_options_generators import get_auth_options
 from infection_monkey.exploit.powershell_utils.credentials import Credentials
 
 CREDENTIALS = [
